Restructured project.
