+++ learning/create_vocabulary_and_vectors.py
