+++ learning/prepare_data.py
@@ -1,0 +1,81 @@
+"""
+Prepares the dataset containing feature vectors and class labels for use by classifiers.
+1)
+Two inputs are loaded using 'load_feature_vectors_and_classes':
+- List of dictionaries of the form:
+    [{feature_vector: [0, 3, 1, ...], num_comments: [10]},
+     {feature_vector: [1, 0, 0, ...], num_comments: [20]]
+- Dict of target classes of the form:
+    {'very_low': {'start': 0, 'end': 10), 'low': {'start': 11, 'end': 20), 'medium': {'start': 21, 'end': 30),
+    'high': {'start': 31, 'end': 40), 'very_high': {'start': 41, 'end': 50),}
+2)
+Based on these, it creates a list of feature vectors and target values using 'get_feature_vectors_and_target_values'.
+With the examples given above, this would look like:
+- [[0, 3, 1, ...], [1, 0, 0, ...]]
+- ['very_low', 'low']
+"""
+import numpy
+
+from pymongo import MongoClient
+from pymongo.collection import Collection
+from pymongo.database import Database
+
+
+def load_feature_vectors_and_classes(db_name):
+    """
+    :param db_name: name of database to use
+    :return:
+        - list of feature vectors
+        - dictionary where the keys are the class labels and the values are dictionaries of the form
+        {start: <integer>, end: <integer>}
+    """
+    print 'Loading feature vectors and target classes...'
+    db = Database(MongoClient(), db_name)
+    collection_names = db.collection_names()
+    if not ('naive_bayes' in collection_names and 'feature_vectors' in collection_names):
+        print 'Database missing collections needed to train classifier on.'
+        exit()
+
+    target_classes = Collection(db, 'naive_bayes').find_one({'type': 'classes'})
+    if 'classes' not in target_classes:
+        raise KeyError("'target_classes' must contain a 'classes' key.")
+
+    feature_vectors = list(Collection(db, 'feature_vectors').find())
+    return feature_vectors, target_classes['classes']
+
+
+def get_feature_vectors_and_target_values(feature_vector_dicts, target_classes):
+    """
+    :param feature_vector_dicts: list of dicts containing feature vectors and number of comments
+    :param target_classes: dictionary containing a 'target class label -> comment interval' mapping
+    :return:
+        - NumPy ndarray containing feature vectors
+        - NumPy ndarray containing target values, with the i-th element corresponding to
+          the target value of the i-th feature vector
+    """
+    if not isinstance(feature_vector_dicts, list):
+        raise TypeError("'feature_vector_dicts' must be a list.")
+    if not isinstance(target_classes, dict):
+        raise TypeError("'target_classes' must be a dict.")
+
+    print 'Preparing %d feature vectors of size %d each...' \
+          % (len(feature_vector_dicts), len(feature_vector_dicts[0]['feature_vector']))
+    feature_vectors = []
+    target_values = []
+
+    def get_class_for_number_of_comments(num_comments):
+        """
+        :param num_comments: number of comments to determine class for
+        :return: class within which the specified number of comments belongs to ('very low', 'high', etc.)
+        """
+        for class_name, comments_range in target_classes.iteritems():
+            if comments_range['start'] <= num_comments <= comments_range['end']:
+                return class_name
+
+    for feature_vector_document in feature_vector_dicts:
+        feature_vectors.append(feature_vector_document['feature_vector'])
+        # Add the class label corresponding to 'feature_vector' to 'target_values'
+        target_values.append(get_class_for_number_of_comments(feature_vector_document['num_comments']))
+
+    target_values_arr = numpy.array(target_values)
+    return numpy.array(feature_vectors), target_values_arr
